Interim commit: Started on named arguments
We need information about the funciton declaration at the call site
now since the compiler needs to emit the named argument names there.

This is also slightly more difficult than i expected since we need to
handle the case where the user supplies a variable pointing to a valid
object.

In that case, we need to extract the correct items from the object. It
is probably not hard, but we prob need to handle it as an edge case.